Remove just-scripts-utils and just-task-logger packages
