--- packages/just-scripts-utils/.npmignore
@@ -1,2 +1,0 @@
-CHANGELOG.*
-__tests__
