--- packages/just-scripts-utils/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "just-scripts-utils",
-  "version": "2.0.0",
-  "description": "Utilities for Just stack scripts",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/microsoft/just"
-  },
-  "main": "./lib/index.js",
-  "scripts": {
-    "build": "tsc",
-    "start": "tsc -w --preserveWatchOutput",
-    "test": "jest",
-    "start-test": "jest --watch"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "engines": {
-    "node": ">=14"
-  },
-  "dependencies": {
-    "fs-extra": "^10.0.0",
-    "just-task-logger": ">=1.2.1 <2.0.0",
-    "semver": "^7.0.0"
-  }
-}
