--- packages/just-scripts-utils/src/logger.ts
@@ -1,1 +1,0 @@
-export * from 'just-task-logger';
