--- packages/just-scripts-utils/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "lib"
-  },
-  "include": ["src"],
-  "exclude": ["**/__tests__/**"]
-}
