+++ packages/just-scripts/package.json
@@ -29,7 +29,6 @@
     "diff-match-patch": "1.0.5",
     "fs-extra": "^10.0.0",
     "glob": "^7.1.3",
-    "just-scripts-utils": "^2.0.0",
     "just-task": ">=1.7.1 <2.0.0",
     "prompts": "^2.4.0",
     "run-parallel-limit": "^1.0.6",
