+++ packages/just-scripts/src/interfaces/PackageJson.ts
