+++ packages/just-scripts/src/tasks/tscTask.ts
@@ -1,6 +1,6 @@
 import * as ts from 'typescript';
 import { resolve, logger, resolveCwd, TaskFunction } from 'just-task';
-import { exec, encodeArgs, spawn } from 'just-scripts-utils';
+import { exec, encodeArgs, spawn } from '../utils';
 import * as fs from 'fs';
 
 export type TscTaskOptions = { [key in keyof ts.CompilerOptions]?: string | boolean | string[] } & {
