+++ packages/just-scripts/src/utils/__tests__/exec.spec.ts
