+++ packages/just-scripts/src/utils/__tests__/paths.spec.ts
