+++ packages/just-scripts/src/utils/__tests__/readPackageJson.spec.ts
