+++ packages/just-scripts/src/utils/index.ts
@@ -1,7 +1,4 @@
 export * from './exec';
-export * from './interfaces/PackageJson';
-export * from './interfaces/RushJson';
-export * from './logger';
 export { mergePackageJson } from './mergePackageJson';
 export * from './paths';
 export * from './readPackageJson';
