+++ packages/just-scripts/src/utils/paths.ts
