--- packages/just-task-logger/.npmignore
@@ -1,2 +1,0 @@
-CHANGELOG.*
-__tests__
