--- packages/just-task-logger/CHANGELOG.md
@@ -1,112 +1,0 @@
-# Change Log - just-task-logger
-
-This log was last generated on Thu, 09 Mar 2023 08:02:09 GMT and should not be manually modified.
-
-<!-- Start content -->
-
-## 1.2.1
-
-Thu, 09 Mar 2023 08:02:09 GMT
-
-### Patches
-
-- Use `yargs-parser` instead of `yargs` to get the `--verbose` argument (elcraig@microsoft.com)
-
-## 1.2.0
-
-Fri, 18 Nov 2022 22:40:12 GMT
-
-### Minor changes
-
-- Require Node 14 (elcraig@microsoft.com)
-
-## 1.1.2
-
-Sat, 10 Sep 2022 00:19:03 GMT
-
-### Patches
-
-- Specify Node 12+ engines requirement (elcraig@microsoft.com)
-
-## 1.1.1
-
-Thu, 14 Jan 2021 20:55:06 GMT
-
-### Patches
-
-- fixing release pipeline and the empty lib folder (kchau@microsoft.com)
-
-## 1.1.0
-
-Wed, 13 Jan 2021 23:22:52 GMT
-
-### Minor changes
-
-- Update typescript and other dependencies (elcraig@microsoft.com)
-
-## 1.0.2
-
-Mon, 02 Nov 2020 16:42:46 GMT
-
-### Patches
-
-- Bump yargs dependency (1581488+christiango@users.noreply.github.com)
-
-## 1.0.1
-
-Mon, 02 Nov 2020 16:26:27 GMT
-
-### Patches
-
-- Publish src folder (1581488+christiango@users.noreply.github.com)
-
-## 1.0.0
-
-Tue, 06 Oct 2020 23:06:25 GMT
-
-### Major changes
-
-- removed just-stack-* (kchau@microsoft.com)
-
-### Patches
-
-- Remove implicit SDI usage (darthtrevino@gmail.com)
-- fixes publish (kchau@microsoft.com)
-
-## 0.3.0
-Thu, 11 Jul 2019 17:08:46 GMT
-
-### Minor changes
-
-- adding perf logging (kchau@microsoft.com)
-
-## 0.2.1
-Mon, 01 Jul 2019 20:15:20 GMT
-
-## 0.2.1
-Wed, 15 May 2019 18:44:04 GMT
-
-### Patches
-
-- Switch microsoft in repo URL to lowercase
-
-## 0.2.0
-Fri, 22 Feb 2019 23:37:36 GMT
-
-### Minor changes
-
-- Consistent interface naming
-
-## 0.1.2
-Fri, 22 Feb 2019 19:03:38 GMT
-
-### Patches
-
-- Update package metadata
-
-## 0.1.1
-Thu, 21 Feb 2019 23:25:53 GMT
-
-### Patches
-
-- Add logger package
