--- packages/just-task-logger/jest.config.js
@@ -1,1 +1,0 @@
-module.exports = require('../../scripts/jest.config');
