--- packages/just-task-logger/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "just-task-logger",
-  "version": "1.2.1",
-  "description": "Logger for Just scripts and tasks",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/microsoft/just"
-  },
-  "main": "./lib/index.js",
-  "scripts": {
-    "build": "tsc",
-    "start": "tsc -w --preserveWatchOutput",
-    "test": "jest",
-    "start-test": "jest --watch"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "engines": {
-    "node": ">=14"
-  },
-  "dependencies": {
-    "chalk": "^4.0.0",
-    "yargs-parser": "^20.2.3"
-  }
-}
