--- packages/just-task-logger/src/index.ts
@@ -1,2 +1,0 @@
-export { logger, Logger } from './logger';
-export { mark } from './perf';
