+++ packages/just-task/package.json
@@ -28,7 +28,6 @@
     "bach": "^2.0.0",
     "chalk": "^4.0.0",
     "fs-extra": "^10.0.0",
-    "just-task-logger": ">=1.2.1 <2.0.0",
     "resolve": "^1.19.0",
     "undertaker": "^1.3.0",
     "undertaker-registry": "^2.0.0",
